+++ CHANGELOG.md
@@ -5,6 +5,7 @@ With this release InfluxDB is moving to Go 1.5.
 ### Features
 - [#5183](https://github.com/influxdb/influxdb/pull/5183): CLI confirms database exists when USE executed. Thanks @pires
 - [#5201](https://github.com/influxdb/influxdb/pull/5201): Allow max UDP buffer size to be configurable. Thanks @sebito91
+- [#5194](https://github.com/influxdb/influxdb/pull/5194): Custom continuous query options per query rather than per node.
 
 ### Bugfixes
 - [#5042](https://github.com/influxdb/influxdb/issues/5042): Count with fill(none) will drop 0 valued intervals.